update examples and tests to use new timeout and idle methods
