+++ fltk/examples/animations.rs
@@ -1,11 +1,4 @@
-use fltk::{
-    app,
-    enums::Color,
-    frame::Frame,
-    image::Pixmap,
-    prelude::*,
-    window::Window,
-};
+use fltk::{app, enums, frame::Frame, image::Pixmap, prelude::*, window::Window};
 
 const PXM: &[&str] = &[
     "50 34 4 1",
@@ -49,32 +42,34 @@ const PXM: &[&str] = &[
     "##################################################",
 ];
 
+fn move_image(mut frm: Frame, handle: app::TimeoutHandle) {
+    let (x, y) = (frm.x(), frm.y());
+    frm.set_pos(x + 5, y);
+    app::redraw();
+    if frm.x() > 260 {
+        app::remove_timeout3(handle)
+    } else {
+        app::repeat_timeout3(0.016, handle);
+    }
+}
+
 fn main() {
     let app = app::App::default();
     let mut wind = Window::default()
-        .with_label("pxm test")
+        .with_label("timeout")
         .with_size(720, 486)
         .center_screen();
-    let mut frame = Frame::new(-30, 200, 30, 30, "");
+    let mut frame = Frame::new(-200, 150, 200, 200, "");
     let mut pxm = Pixmap::new(PXM).unwrap();
     pxm.scale(200, 200, true, true);
-    frame.set_image(Some(pxm));
-    wind.set_color(Color::White);
+    frame.set_image_scaled(Some(pxm));
+    wind.set_color(enums::Color::White);
     wind.end();
     wind.show_with_env_args();
 
-    app::add_idle(move || {
-        let x = frame.x();
-        let y = frame.y();
-        let w = frame.width();
-        let h = frame.height();
-        if x > wind.width() + w + 30 {
-            app.quit();
-        }
-        frame.resize(x + 5, y, w, h);
-        wind.redraw();
-        app::sleep(0.016); // sleeps are necessary when calling redraw in the event loop
+    app::add_timeout3(0.016, move |handle| {
+        let frame = frame.clone();
+        move_image(frame, handle);
     });
-
     app.run().unwrap();
 }
